Manually extracted every SA graphic
All of these can be compressed with lz77 type 10 and then by huffman 4-bit mode, the ones that say in the filename 'HUFFMAN28' are stored in the ROM with JUST huffman 8-bit mode, no lz77.

The numbers at the start of very filename are important, as they can be used to jump directly to the pointer stored on the table in the ROM at 0x26042C.

There might be some unused content here.