Added the tilemap for the titlescreen
Tilemap Studio seems to work well for adjusting the tilemap, since the logo swaps between 2 palettes, one can quickly change all the tiles that use palette 2 to palette 1.

But of course the best case scenario would be to cleverly make use of both palettes to create a high quality logo.